Buy crypto feature

+++ src/main/java/com/example/ctsusermanagement/controller/PortfolioController.java
@@ -6,12 +6,9 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.math.BigDecimal;
 
 @Tag(name = "Portfolio", description = "Portfolio API")
 @RestController
@@ -24,7 +21,7 @@ public class PortfolioController {
     @SecurityRequirement(name = "TOKEN")
     @Operation(summary = "Deposit")
     @PostMapping("/deposit")
-    public ResponseEntity<String> deposit(BigDecimal amount) {
+    public ResponseEntity<String> deposit(Double amount) {
         portfolioService.deposit(amount);
         return ResponseEntity.ok("Deposit successfully accepted");
     }
