+++ src/main/java/com/example/ctsusermanagement/service/PortfolioService.java
@@ -1,8 +1,6 @@
 package com.example.ctsusermanagement.service;
 
-import java.math.BigDecimal;
-
 public interface PortfolioService {
 
-    void deposit(BigDecimal deposit);
+    void deposit(Double deposit);
 }
